+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # solana
 
+## 2.2.1
+
+### Patch Changes
+
+- refactor default installing all tools
+
+## 2.2.0
+
+### Minor Changes
+
+- added installer for trident
+
+### Patch Changes
+
+- fix install spinners and messages
+
 ## 2.1.0
 
 ### Minor Changes